Improve provider validation pre-commit
The provider validation only checked if classes and packages were
belonging to the right package, but they did not check if the classes or
packages specified could actually be imported - which could lead to
webserver crash as in case of #28393.

That requires the pre-commit code to be moved to the Breeze container,
because only there we have all the packages installed that are needed
to import the packages and classes.

We have to also make an exclusion for MSSQL and MySQL classes because
they are missing for ARM images, so on ARM import errors for those
are expected.

Fixes: #28510
